Send forward URL in webhook responses
